split(store/sqlite): split store_authz_test.go into topic-focused test files
Replace the 2292-line monolithic store_authz_test.go with seven files,
each under the 500-line limit set by T65:

- store_authz_masks_test.go       – EffectiveMask, mask helpers
- store_authz_user_listing_test.go  – UserAuthzResourcesList
- store_authz_group_listing_test.go – GroupAuthzResourcesList
- store_authz_membership_test.go  – add/remove/grant/revoke operations
- store_authz_restrict_delete_test.go – restrict-delete FK enforcement
- store_authz_resource_users_test.go  – ResourceAuthzUsersList
- store_authz_resource_groups_test.go – ResourceAuthzGroupsList

Imports are alphabetically ordered in every new file (stdlib then
third-party), consistent with the rest of the package.
